Move auditor interface to point of consumption
Move auditor interface to point of consumption. Also rename auditor
implementation to match its design.
